wip: implement corrections feedback ui

+++ api/src/main/java/au/org/aodn/nrmn/restapi/dto/correction/CorrectionRowPutDto.java
@@ -12,7 +12,7 @@ public class CorrectionRowPutDto {
     Long id;
     Collection<Integer> observationIds;
     Long surveyId;
-    Long diverId;
+    Integer diverId;
     String initials;
     String siteCode;
     Integer depth;
